# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PV DataStore API — Puerto Vallarta data aggregation platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sources/{id}/sync:
    post:
      operationId: syncDataSource
      tags: [sources]
      summary: Trigger a data source sync
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sync triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
components:
  schemas:
...
    SyncResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        sourceId:
          type: integer
        recordsProcessed:
          type: integer
      required:
        - success
        - message
        - sourceId