# lib/api-zod/src/generated/types/syncResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PV DataStore API — Puerto Vallarta data aggregation platform
 * OpenAPI spec version: 0.1.0
 */

export interface SyncResult {
  success: boolean;
  message: string;
  sourceId: number;
  recordsProcessed?: number;
}